refactor: move Graph module to collections directory
- Create collections directory structure
- Move Graph.ts to collections/Graph.ts
- Update import paths to use ../
- Update index.ts export to use collections path
- All tests continue to pass
